requirements.txt
annotated-types==0.7.0
anyio==4.12.1
attrs==25.4.0
Automat==25.4.16
beautifulsoup4==4.14.3
bs4==0.0.2
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.5
click==8.3.1
constantly==23.10.4
cryptography==46.0.5
cssselect==1.4.0
defusedxml==0.7.1
filelock==3.25.2
greenlet==3.3.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
hyperlink==21.0.0
idna==3.11
Incremental==24.11.0
itemadapter==0.13.1
itemloaders==1.4.0
jmespath==1.1.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
lxml==6.0.2
mcp==1.26.0
packaging==26.0
parsel==1.11.0
playwright==1.58.0
Protego==0.6.0
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
PyDispatcher==2.0.7
pyee==13.0.1
PyJWT==2.11.0
pyOpenSSL==26.0.0
python-dotenv==1.2.2
python-multipart==0.0.22
queuelib==1.9.0
referencing==0.37.0
requests==2.32.5
requests-file==3.0.1
rpds-py==0.30.0
Scrapy==2.14.2
service-identity==24.2.0
soupsieve==2.8.3
sse-starlette==3.3.2
starlette==0.52.1
tldextract==5.3.1
Twisted==25.5.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.41.0
w3lib==2.4.1
zope.interface==8.2
